// types/Subscription.ts
import { DocumentData, DocumentReference, Timestamp } from 'firebase-admin/firestore'
import Stripe from 'stripe'

export interface Subscription {
  id?: string;
  metadata: {
    [name: string]: string;
  },
  stripeLink: string
  role: string | null;
  quantity: number;
  items: Stripe.SubscriptionItem[];
  product: DocumentReference<DocumentData>
  price: DocumentReference<DocumentData>
  prices: Array<DocumentReference<DocumentData>>
  payment_method?: string;
  latest_invoice?: string;
  status:
  | 'active'
  | 'canceled'
  | 'incomplete'
  | 'incomplete_expired'
  | 'past_due'
  | 'trialing'
  | 'unpaid';

  cancel_at_period_end: boolean;
  created: Timestamp
  current_period_start: Timestamp
  cuurent_period_end: Timestamp
  canceled_at: Timestamp | null
  trial_start: Timestamp | null
  trial_end: Timestamp | null
}